fixed all typing errors and simplified maps and lists
