+++ .github/workflows/static-analysis.yml
@@ -7,24 +7,33 @@ on:
   pull_request:
     branches:
       - main
-
 jobs:
   php-cs-fixer:
     name: "Lint & Analyse"
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2
+
+      # Setup the correct PHP version globally
+      - name: Setup PHP
+        uses: shivammathur/setup-php@v2
+        with:
+          php-version: '8.3.16'
+
       - name: Cache Composer dependencies
         uses: actions/cache@v2
         with:
           path: /tmp/composer-cache
           key: ${{ runner.os }}-${{ hashFiles('**/composer.lock') }}
+
       - uses: php-actions/composer@v6
         with:
           progress: yes
-          php_version: 8.3
+          php_version: 8.3.16
           version: 2
+
       - name: "PHP-CS-Fixer"
         run: vendor/bin/php-cs-fixer fix --dry-run
+
       - name: "PSalm"
         run: vendor/bin/psalm --show-info=true
