+++ .github/workflows/unit-test.yml
@@ -14,7 +14,7 @@ jobs:
     name: "Running Unit Tests"
     strategy:
         matrix:
-            php: ['8.1', '8.2', '8.3']
+            php: ['8.1.31', '8.2.27', '8.3.16']
 
     steps:
       - uses: actions/checkout@v2
