+++ composer.json
@@ -48,9 +48,9 @@
         "nyholm/psr7": "^1.3",
         "nyholm/psr7-server": "^1.0",
         "kriswallsmith/buzz": "^1.2",
-        "vimeo/psalm": "^5.0",
+        "vimeo/psalm": "^6.5",
         "friendsofphp/php-cs-fixer": "3.68.5",
-        "psalm/plugin-phpunit": "^0.18",
+        "psalm/plugin-phpunit": "^0.19",
         "monolog/monolog": "^2.2",
         "symfony/uid": "^5.0",
         "symfony/var-dumper": "^5.0",
