+++ src/Bolt/BoltResult.php
@@ -26,7 +26,7 @@ use Laudis\Neo4j\Formatter\SummarizedResultFormatter;
 /**
  * @psalm-import-type BoltCypherStats from SummarizedResultFormatter
  *
- * @implements Iterator<int, list>
+ * @implements Iterator<int, list<mixed>>
  */
 final class BoltResult implements Iterator
 {
@@ -71,7 +71,7 @@ final class BoltResult implements Iterator
     }
 
     /**
-     * @return Generator<int, list>
+     * @return Generator<int, list<mixed>>
      */
     public function iterator(): Generator
     {
@@ -113,10 +113,15 @@ final class BoltResult implements Iterator
     }
 
     /**
-     * @return list
+     * @psalm-suppress InvalidNullableReturnType
+     *
+     * @return list<mixed>
      */
     public function current(): array
     {
+        /**
+         * @psalm-suppress NullableReturnStatement
+         */
         return $this->getIt()->current();
     }
